Support for add medicatons in the discpense list page
