+++ public/locale/en.json
@@ -424,6 +424,7 @@
   "add_diagnosis_one": "Add Diagnosis",
   "add_diagnosis_other": "Add another Diagnosis",
   "add_discount": "Add Discount",
+  "add_dosage_and_dispense": "Add Dosage & Dispense",
   "add_drawings": "Add Drawings",
   "add_edit_interpretation": "Add/Edit Interpretation",
   "add_element": "Add element",
@@ -998,6 +999,9 @@
   "cancel": "Cancel",
   "cancel_appointment": "Cancel Appointment",
   "cancel_appointment_warning": "This action cannot be undone. The appointment will be cancelled and the patient will be notified.",
+  "cancel_dispense": "Cancel Dispense",
+  "cancel_dispense_confirmation": "Are you sure you want to cancel this dispense? This action cannot be undone.",
+  "cancel_dispense_with_invoices_confirmation": "Are you sure you want to cancel this dispense? This will also cancel {{count}} associated invoice(s). This action cannot be undone.",
   "cancel_invoice": "Cancel Invoice",
   "cancel_invoice_confirmation": "Are you sure you want to cancel invoice {{invoiceNumber}}? This action cannot be undone and the charge items will become unbilled.",
   "cancel_payment": "Cancel Payment",
@@ -1821,6 +1825,7 @@
   "dispatched_qty": "Dispatched Qty",
   "dispatched_quantity": "Dispatched Quantity",
   "dispense": "Dispense",
+  "dispense_and_invoices_cancelled": "Dispense and invoices cancelled successfully",
   "dispense_for_patient": "Create dispense for {{patientName}}",
   "dispense_history": "Dispense History",
   "dispense_medications": "Dispense Medications",
@@ -2250,6 +2255,7 @@
   "error_calculating_price": "Error calculating price",
   "error_completing_medication_return_items": "Error completing medication return items",
   "error_creating_consent": "Error creating consent",
+  "error_creating_medication_dispense": "Error creating medication dispense",
   "error_creating_medication_return": "Error creating medication return",
   "error_creating_product": "Error creating product",
   "error_creating_supply_delivery": "Error creating supply delivery",
@@ -2379,6 +2385,7 @@
   "failed_to_add_to_template": "Failed to add medication to template",
   "failed_to_apply_template": "Failed to apply template",
   "failed_to_archive_child_tag": "Failed to archive child tag",
+  "failed_to_cancel_dispense": "Failed to cancel dispense",
   "failed_to_cancel_invoice": "Failed to cancel invoice",
   "failed_to_cancel_payment": "Failed to cancel payment",
   "failed_to_capture_image": "Failed to capture image",
@@ -3173,6 +3180,7 @@
   "medication_administration_saved": "Medicine Administration saved",
   "medication_already_marked_as_error": "Medication already marked as entered in error",
   "medication_dispense": "Medication Dispense",
+  "medication_dispense_created": "Medication dispense created successfully",
   "medication_dispense_details": "Medication Dispense Details",
   "medication_dispense_invoice_alert": "Dispense Medicine now",
   "medication_dispense_status__cancelled": "Cancelled",
@@ -3359,6 +3367,8 @@
   "no_accounts_found": "No accounts found",
   "no_active_account_found": "No active account found",
   "no_active_beds_assigned": "No active beds assigned",
+  "no_active_encounter_description": "An active encounter is required to add a medication dispense. Please create an encounter first.",
+  "no_active_encounter_for_patient": "No active encounter found for this patient",
   "no_active_encounters_found": "No active encounters found",
   "no_active_medication_recorded": "No Active Medication Recorded",
   "no_active_products": "No active products",
@@ -4108,6 +4118,7 @@
   "please_fix_errors": "Please fix the errors in the highlighted fields and try submitting again.",
   "please_select_a_consent": "Please select a consent",
   "please_select_a_facility": "Please select a facility",
+  "please_select_a_lot": "Please select a lot",
   "please_select_at_least_one_delivery": "Please select at least one delivery",
   "please_select_at_least_one_item": "Please select at least one item",
   "please_select_breathlessness_level": "Please select Breathlessness Level",
@@ -5162,6 +5173,7 @@
   "select_status": "Select Status",
   "select_stock": "Select stock",
   "select_stock_at_row": "Please select a stock item at row {{row}}",
+  "select_stock_lot": "Select Stock Lot",
   "select_stroke": "Select Stroke",
   "select_sub_department": "Select sub-department",
   "select_subject_type": "Select Subject Type",
@@ -6189,6 +6201,7 @@
   "years_of_experience": "Years of Experience",
   "years_of_experience_of_the_doctor": "Years of Experience of the Doctor",
   "yes": "Yes",
+  "yes_cancel": "Yes, Cancel",
   "yes_cancel_invoice": "Yes, Cancel Invoice",
   "yes_cancel_payment": "Yes, Cancel Payment",
   "yesterday": "Yesterday",
